requirements.txt
aiogram==2.25.2
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
Babel==2.9.1
certifi==2024.2.2
charset-normalizer==3.3.2
et-xmlfile==1.1.0
frozenlist==1.4.1
idna==3.6
magic-filter==1.0.12
multidict==6.0.5
numexpr==2.10.0
numpy==1.26.4
openpyxl==3.1.2
pandas==2.2.1
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
tzdata==2024.1
yarl==1.9.4
